requirements.txt
numpy
scipy
matplotlib
seaborn
pandas
xarray
rasterio
pyproj
torch
torchvision
pyyaml
tqdm
hydra-core
scikit-image
scikit-learn
